refactor(products): move product comment routes to separate file
